functions/requirements.txt
firebase_functions~=0.1.0
line-bot-sdk
flask
openai >= 1.0.0
gunicorn
firebase-admin
python-dotenv